+++ extend/bajo/hook/dobo.sumba-site@after-create-record.js
@@ -1,0 +1,8 @@
+import { createRefs, getAllFixtures } from '../../../lib/create-new-site.js'
+
+async function afterCreateRecord (body, result, options) {
+  const fixtures = await getAllFixtures.call(this, result.data.alias)
+  await createRefs.call(this, result.data, fixtures, options)
+}
+
+export default afterCreateRecord