+++ extend/bajo/hook/dobo.sumba-site@after-remove-record.js
@@ -1,0 +1,7 @@
+import { removeRefs } from '../../../lib/remove-site.js'
+
+async function afterRemoveRecord (id, result, options) {
+  await removeRefs.call(this, result.oldData, options)
+}
+
+export default afterRemoveRecord