+++ extend/bajo/hook/dobo.sumba-user@before-create-record.js
@@ -1,7 +1,5 @@
-import resetToken from '../../../lib/reset-token.js'
-
 async function beforeCreateRecord (body, options = {}) {
-  const { token, salt } = await resetToken.call(this)
+  const { token, salt } = await this.resetToken()
   body.token = token
   body.salt = salt
 }