+++ extend/bajo/hook/dobo.sumba-user@before-record-validation.js
@@ -1,8 +1,6 @@
-import passwordRule from '../../../lib/password-rule.js'
-
 async function beforeRecordValidation (body, options = {}) {
   const { set } = this.app.lib._
-  const password = await passwordRule.call(this, options.req)
+  const password = await this.passwordRule(options.req)
   const rule = { password }
   set(options, 'validation.params.rule', rule)
 }