+++ extend/bajo/hook/waibu-rest-api@pre-parsing.js
@@ -1,10 +1,11 @@
-import { checkUserId, checkTeam } from '../../../lib/util.js'
+import { checkUserId, checkTeam, checkinterSite } from '../../../lib/util.js'
 
 const preParsing = {
   level: 10,
   handler: async function (req, reply) {
     await checkUserId.call(this, req, reply, 'waibuRestApi')
     await checkTeam.call(this, req, reply, 'waibuRestApi')
+    await checkinterSite.call(this, req, reply)
   }
 }