+++ extend/waibuMpa/extend/waibuAdmin/route/_is_/cache/@action.js
@@ -1,10 +1,12 @@
 const action = {
   method: ['GET', 'POST'],
-  title: 'userSession',
+  title: 'cacheStorage',
+  interSite: true,
   handler: async function (req, reply) {
+    if (!this.app.bajoCache) throw this.error('_notFound')
     const { importModule } = this.app.bajo
     const crudSkel = await importModule('waibuAdmin:/lib/crud-skel.js')
-    return await crudSkel.call(this, 'WmpaSession', req, reply)
+    return await crudSkel.call(this, 'CacheStorage', req, reply)
   }
 }