+++ extend/waibuMpa/extend/waibuAdmin/route/_is_/session/@action.js
@@ -1,0 +1,12 @@
+const action = {
+  method: ['GET', 'POST'],
+  title: 'userSession',
+  interSite: true,
+  handler: async function (req, reply) {
+    const { importModule } = this.app.bajo
+    const crudSkel = await importModule('waibuAdmin:/lib/crud-skel.js')
+    return await crudSkel.call(this, 'WmpaSession', req, reply)
+  }
+}
+
+export default action