+++ extend/waibuMpa/extend/waibuAdmin/route/_is_/site/@action.js
@@ -1,11 +1,11 @@
 const action = {
   method: ['GET', 'POST'],
-  title: 'cacheStorage',
+  title: 'manageAllSite',
+  interSite: true,
   handler: async function (req, reply) {
-    if (!this.app.bajoCache) throw this.error('_notFound')
     const { importModule } = this.app.bajo
     const crudSkel = await importModule('waibuAdmin:/lib/crud-skel.js')
-    return await crudSkel.call(this, 'CacheStorage', req, reply)
+    return await crudSkel.call(this, 'SumbaSite', req, reply)
   }
 }